fix(macros): remove url-resolver feature flag, gate on platform instead
Remove the `url-resolver` feature flag entirely and make URL resolution
always available on native (non-WASM) targets. The `native` cfg alias
(from cfg_aliases) does not propagate to consuming crates, so
macro-generated code now uses the raw platform check
`cfg(not(all(target_family = "wasm", target_os = "unknown")))` instead.

Library source code retains `cfg(native)` since those crates have their
own build.rs with cfg_aliases.

Restructure installed_apps tests into separate compilation units to
avoid `#[macro_export]` name collisions from multiple invocations.

Fixes #3510
