+++ crates/reinhardt-core/macros/tests/installed_apps_tests.rs
@@ -1,65 +1,41 @@
 //! Tests for the installed_apps! macro
+//!
+//! Uses a single `installed_apps!` invocation since the macro emits a
+//! `#[macro_export]` helper (`__reinhardt_for_each_app`) that must be
+//! unique at crate scope.
 
 use reinhardt_macros::installed_apps;
 
-#[test]
-fn test_installed_apps_empty() {
-	installed_apps! {}
-
-	// Arrange
-	use std::str::FromStr;
-
-	// Act
-	let apps = InstalledApp::all_apps();
-	let from_str_result = InstalledApp::from_str("anything");
-
-	// Assert
-	assert_eq!(apps, Vec::<String>::new());
-	assert!(from_str_result.is_err());
-	assert_eq!(from_str_result.unwrap_err(), "Unknown app: anything");
+installed_apps! {
+	auth: "myproject.auth",
+	sessions: "myproject.sessions",
+	contenttypes: "myproject.contenttypes",
+	myapp: "apps.myapp",
+	another: "custom.another",
 }
 
 #[test]
 fn test_installed_apps_basic() {
-	installed_apps! {
-		auth: "myproject.auth",
-		contenttypes: "myproject.contenttypes",
-	}
-
 	let apps = InstalledApp::all_apps();
-	assert_eq!(apps.len(), 2);
+	assert_eq!(apps.len(), 5);
 	assert!(apps.contains(&"myproject.auth".to_string()));
 	assert!(apps.contains(&"myproject.contenttypes".to_string()));
 }
 
 #[test]
 fn test_installed_apps_enum() {
-	installed_apps! {
-		auth: "myproject.auth",
-		sessions: "myproject.sessions",
-	}
-
 	assert_eq!(InstalledApp::auth.path(), "myproject.auth");
 	assert_eq!(InstalledApp::sessions.path(), "myproject.sessions");
 }
 
 #[test]
 fn test_installed_apps_display() {
-	installed_apps! {
-		auth: "myproject.auth",
-	}
-
 	let app = InstalledApp::auth;
 	assert_eq!(format!("{}", app), "myproject.auth");
 }
 
 #[test]
 fn test_installed_apps_from_str() {
-	installed_apps! {
-		auth: "myproject.auth",
-		sessions: "myproject.sessions",
-	}
-
 	use std::str::FromStr;
 
 	let auth = InstalledApp::from_str("myproject.auth");
@@ -72,35 +48,13 @@ fn test_installed_apps_from_str() {
 
 #[test]
 fn test_installed_apps_with_user_apps() {
-	installed_apps! {
-		auth: "myproject.auth",
-		myapp: "apps.myapp",
-		another: "custom.another",
-	}
-
 	let apps = InstalledApp::all_apps();
-	assert_eq!(apps.len(), 3);
 	assert!(apps.contains(&"apps.myapp".to_string()));
 	assert!(apps.contains(&"custom.another".to_string()));
 }
 
 #[test]
-fn test_installed_apps_single_app() {
-	installed_apps! {
-		auth: "myproject.auth",
-	}
-
-	let apps = InstalledApp::all_apps();
-	assert_eq!(apps.len(), 1);
-}
-
-#[test]
 fn test_installed_apps_equality() {
-	installed_apps! {
-		auth: "myproject.auth",
-		sessions: "myproject.sessions",
-	}
-
 	let app1 = InstalledApp::auth;
 	let app2 = InstalledApp::auth;
 	let app3 = InstalledApp::sessions;
@@ -111,10 +65,6 @@ fn test_installed_apps_equality() {
 
 #[test]
 fn test_installed_apps_debug() {
-	installed_apps! {
-		auth: "myproject.auth",
-	}
-
 	let app = InstalledApp::auth;
 	let debug_str = format!("{:?}", app);
 	assert!(debug_str.contains("auth"));
