+++ crates/reinhardt-urls/Cargo.toml
@@ -51,10 +51,9 @@ routers-macros = ["dep:reinhardt-routers-macros"]
 proxy = []
 database = []
 client-router = ["reinhardt-core/reactive", "routers"]
-urls-full = ["routers", "routers-macros", "proxy", "url-resolver"]
+urls-full = ["routers", "routers-macros", "proxy"]
 full = ["database", "unified-router"]
 unified-router = []
-url-resolver = []
 
 [build-dependencies]
 cfg_aliases = "0.2"
