+++ examples/Cargo.toml
@@ -40,13 +40,12 @@ chrono = { version = "0.4.42", features = ["serde"] }
 anyhow = "1.0.100"
 
 # Shared lint configuration for all examples.
-# reinhardt_macros generates code with cfg(native), cfg(wasm), and
-# cfg(feature = "url-resolver") that are defined in the main workspace but not
-# in the examples workspace. Declare them here so check-cfg does not reject them.
+# reinhardt_macros generates code with cfg(native), cfg(wasm) that are defined
+# in the main workspace but not in the examples workspace. Declare them here so
+# check-cfg does not reject them.
 [workspace.lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = [
 	'cfg(native)',
 	'cfg(wasm)',
 	'cfg(with_reinhardt)',
-	'cfg(feature, values("url-resolver"))',
 ] }
